+++ src/main/java/factor/Arrayer.java
@@ -1,4 +1,4 @@
-package xfactor;
+package factor;
 
 import java.util.Arrays;
